pyproject.toml
# pyproject.toml

[build-system]
requires      = ["setuptools>=61.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "alphaiq-sdk"
version = "0.0.1"
description = "Access AlphaIQ quantitative linguistics SPINDEX risk scores"
readme = "README.md"
authors = [{ name = "AlphaIQ", email = "contact@alphaiq.ai" }]

classifiers = [
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
]
keywords = ["risk", "AI", "fintech", "investing","NLP"]
dependencies = [
    "requests",
    "python-dotenv",
    "uplink"
]
requires-python = ">=3.7"

[project.scripts]
alphaiq_sdk = "alphaiq_sdk.__main__:main"